backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
openai-whisper                # or openai-whisper
torch                  # CPU or your CUDA build
